+++ docs/scenes/get-started/set-up-your-environment/SelectCard.tsx
@@ -27,13 +27,13 @@ export function SelectCard({
     <ButtonBase onClick={onClick}>
       <div
         className={mergeClasses(
-          'flex w-[250px] flex-col overflow-hidden rounded-lg border border-default shadow-xs transition-all',
+          'border-default flex w-[250px] flex-col overflow-hidden rounded-lg border shadow-xs transition-all',
           'hocus:scale-[102%] hocus:shadow-sm'
         )}>
         <div
           className={mergeClasses(
-            'border-b border-default',
-            isSelected ? 'bg-gradient-to-b from-palette-blue3 to-palette-blue4' : 'bg-subtle'
+            'border-default border-b',
+            isSelected ? 'from-palette-blue3 to-palette-blue4 bg-linear-to-b' : 'bg-subtle'
           )}>
           <picture className="relative flex h-full w-auto items-end">
             {themeName !== 'light' && (
@@ -55,13 +55,13 @@ export function SelectCard({
             <div className="relative">
               <div
                 className={mergeClasses(
-                  'size-5 rounded-full border bg-default',
+                  'bg-default size-5 rounded-full border',
                   isSelected ? 'border-palette-blue9' : 'border-default'
                 )}
               />
               <div
                 className={mergeClasses(
-                  'absolute right-1 top-1 size-3 rounded-full',
+                  'absolute top-1 right-1 size-3 rounded-full',
                   isSelected ? 'border-palette-blue9 bg-palette-blue9' : 'bg-transparent'
                 )}
               />