+++ packages/expo-maps/ios/AppleMapsViewState.swift
@@ -11,6 +11,9 @@ public class AppleMapsViewiOS18State: ObservableObject {
   @Published var lookAroundScene: MKLookAroundScene?
   @Published var lookAroundPresented: Bool = false
   var hasInitializedCamera: Bool = false
+  var lastKnownDistance: Double?
+  var lastKnownHeading: Double = 0
+  var lastKnownPitch: Double = 0
 }
 
 @available(iOS 17.0, *)